code refactor: move some AMDS server side codes into subfolder
