+++ source/Connection/AMDSThreadedTCPDataServer.h
@@ -4,7 +4,7 @@
 #include <QObject>
 #include <QThread>
 
-#include "source/AMDSTCPDataServer.h"
+#include "source/Connection/AMDSTCPDataServer.h"
 
 /**
   * Class which maintains the TCPServer in a separate thread, and passes signal messages to it.
